requirements.txt
Flask==2.0.1
redis==3.5.3
Pillow==8.3.2
python-magic==0.4.24
APScheduler==3.8.1
Flask-Babel==2.0.0
gunicorn==20.1.0
psutil==5.8.0
werkzeug==2.0.1